// src/app/interfaces/bid-management/bid-management.interface.ts
export interface IBidDocument {
  url: string;
  extension: string;
  type: string;
  name: string;
}

interface IPreBiddingMeetingOnSite {
  isChecked?: boolean;
  type: 'Onsite';
  location?: string;
  date?: string;
  time?: string;
  instruction?: string;
  isMandatory?: boolean;
}

interface IPreBiddingMeetingOnline {
  isChecked?: boolean;
  type: 'Online';
  meeting?: IMeeting | string;
  isMandatory?: boolean;
}

interface ISiteWalkthrough {
  isChecked?: boolean;
  location?: string;
  date?: string;
  time?: string;
  instruction?: string;
  isMandatory?: boolean;
}

interface IRFIDeadline {
  date?: string;
  time?: string;
  isChecked?: boolean;
}

export interface IBidManagement {
  user: string | IUserInterface;
  projectName: string;
  address: string;
  zipCode: string;
  city: string;
  state: string;
  country: string;
  constructionTypes: string[];
  projectType: string[];
  projectBuildingUse: string[];
  stage: string;
  estimatedStartDate: string;
  bidDueDate: string;
  estimatedCompletionDate: string;
  squareFootage: string;
  projectValue: number;
  estimatedDuration: string;
  durationType: 'days' | 'years' | 'months';
  description: string;
  specialInstructions: string;
  teamMembers: IBidManagementProjectTeamMember[];
  selectedTrades: ITrade[] | string[];
  projectFiles: IBidDocument[];
  selectedTeamMembers: string[];
  status: 'draft' | 'expired' | 'active' | 'archived' | 'bid closed';
  _id: string;
  createdAt: string;
  updatedAt: string;

  platformType: 'Public' | 'Private';
  isMatchingWithTrades: boolean;
  invitedMembers: string[];
  biddingTeam?: any | any[];
  invitedMembersAssets: {
    name: string;
    url: string;
    extension: string;
    type: string;
  }[];

  preBiddingMeeting?: IPreBiddingMeetingOnSite | IPreBiddingMeetingOnline;
  siteWalkthrough?: ISiteWalkthrough;
  rfiDeadline?: IRFIDeadline;

  sendInMyNetwork?: boolean;

  networkUsers?: {
    userId: string;
    email: string;
  }[];

  currency?: CurrencyType;
}

export interface IBidManagementResponse {
  paginationInfo: {
    currentPage: number;
    pages: number;
    totalRecords: number;
    perPage: number;
  };
  records: (IBidManagement & { userDetails: IUserInterface[] })[];
}

export interface IBidManagementProjectTeamMember {
  user: string;
  name: string;
  role: string;
  companyName: string;
  location: string;
  phoneNumber: string;
  email: string;
  _id: string;
  createdAt: string;
  updatedAt: string;
}
// ...
export interface IBidProjectReport {
  _id: string;
  reasonType: string[];
  reason: string;
  project: string | IBidManagement;
  user: string | IUserInterface;
  createdAt: string;
  updatedAt: string;
}